requirements.txt
azure-mgmt-resource
azure-mgmt-web
azure-mgmt-consumption
azure-identity
